refactor to reduce complexity
